Record objects

When read_ctl() reads a control stream, it returns an nmrec_ctl_records S3 object. Its records field is a list of R6 objects that inherits from nmrec_record. When parsing within a record is supported for a given record type (e.g., "estimation"), the child class is specific for that type (e.g., nmrec_record_estimation). On the other hand, records for which parsing has not been implemented have a type nmrec_record_raw.

Details

Fields

name: the standardized name for the record (e.g., "estimation").

values: a list of capturing the record's content. This is constructed by the $parse() method and consists of "elements" like white space and line breaks interspersed with option objects.

Methods

parse: parse the record's lines into values, enabling inspection and modification.

get_options: return list of nmrec_option objects from values.

get_lines: return original lines passed during initialization. These values are never updated.

format: render the record as a string. If the record has been parsed, this is derived from values.

See also